requirements.txt
streamlit
azure-cognitiveservices-speech
soundfile
sounddevice
scipy
